+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 5.2.0-beta-2 - 2025-05-22
+
 ### Added
 
 - AI Guard: config APIs.